backend/pyproject.toml
[tool.poetry]
name = "invoice-processor"
version = "0.1.0"
description = "OCR-powered invoice data extraction"
authors = []
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
psycopg2-binary = "^2.9"
python-jose = {extras = ["cryptography"], version = "^3.3"}
httpx = "^0.27"
python-dotenv = "^1.0"
azure-storage-blob = "^12.19"
python-multipart = "^0.0.9"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
ruff = "^0.4"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
